Add cache_mode option to gclient.
gclient supports the cache_dir option, which causes it to clone repositories from a local mirror with the --shared option. The problem with this is that it breaks workflows such as git push and git fetch, which interact with the local mirror rather than the remote. This introduces a new cache_mode="reference" option which causes gclient to clone repositories with --reference <mirror>